moved logic to app.jsx to streamline components
